+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.1.0-beta.1 (Unreleased)
 ### New Features
+- Added support for pre-built business card recognition.
+- Added support for pre-built invoices recognition.
 - Added implementation support to create a composed model from the `FormTrainingClient` by calling method `beginCreateComposedModel`.
 - Added properties `modelName` and `customFormModelProperties` to types `CustomFormModel` and `CustomFormModelInfo`.
 - Added property `modelName` to `TrainingOptions` and new type `CreateComposedModelOptions`.
@@ -9,7 +11,6 @@
 - Added properties `modelId` and `formTypeConfidence` to `RecognizedForm`.
 - Added support for providing locale info when recognizing receipts and business cards.
 Supported locales include support EN-US, EN-AU, EN-CA, EN-GB, EN-IN.
-- Added support for pre-built business card recognition.
 - Added support to `beginRecognizeContent` to recognize selection marks such as check boxes and radio buttons.
 - Added support to train and recognize custom forms with selection marks such as check boxes and radio buttons.
 This functionality is only available in trained with labels scenarios.